[REF] base_automation: move some static files
Move some file up a folder:
- ./static/src/js/... => ./static/src/...
- ./static/src/xml/... => ./static/src/...

Task-3085360
