+++ addons/base_automation/static/src/base_automation_error_dialog.js
