+++ addons/base_automation/static/tests/base_automation_error_dialog.js
@@ -15,9 +15,8 @@ import { MainComponentsContainer } from "@web/core/main_components_container";
 import { notificationService } from "@web/core/notifications/notification_service";
 import { registry } from "@web/core/registry";
 import { uiService } from "@web/core/ui/ui_service";
-
 import { getFixture, mount, nextTick, patchWithCleanup } from "@web/../tests/helpers/utils";
-import { BaseAutomationErrorDialog } from "../src/js/base_automation_error_dialog";
+import { BaseAutomationErrorDialog } from "@base_automation/base_automation_error_dialog";
 
 const { toRaw } = owl;
